refactor: reorganize files into proper package structure
