+++ src/chronicler/pipeline/__init__.py
@@ -1,0 +1,9 @@
+from .frames import Frame, TextFrame, ImageFrame, DocumentFrame
+from .pipeline import BaseTransport, BaseProcessor, Pipeline
+from .pipecat_runner import run_pipecat
+
+__all__ = [
+    'Frame', 'TextFrame', 'ImageFrame', 'DocumentFrame',
+    'BaseTransport', 'BaseProcessor', 'Pipeline',
+    'run_pipecat'
+]
